Ejercicio 10 CORRECT in juez!

+++ Grafos/Ejercicio10.cpp
@@ -10,7 +10,7 @@
 
 bool resuelveCaso() 
 {
-    int V, A, v, w;
+    size_t V, A, v, w;
     std::cin >> V; // número de vértices
     if (!std::cin)  // hemos terminado de procesar todos los casos
         return false;
